feat: merge enums into one file
