+++ packages/core/src/abstract-dialect/query-generator-typescript.ts
@@ -5,6 +5,7 @@ import { randomUUID } from 'node:crypto';
 import NodeUtil from 'node:util';
 import type { Class } from 'type-fest';
 import { ConstraintChecking } from '../deferrable.js';
+import { IndexHints, TableHints } from '../enums.js';
 import { AssociationPath } from '../expression-builders/association-path.js';
 import { Attribute } from '../expression-builders/attribute.js';
 import { BaseSqlExpression } from '../expression-builders/base-sql-expression.js';
@@ -18,14 +19,12 @@ import { List } from '../expression-builders/list.js';
 import { Literal } from '../expression-builders/literal.js';
 import { Value } from '../expression-builders/value.js';
 import { Where } from '../expression-builders/where.js';
-import { IndexHints } from '../index-hints.js';
 import type { ModelDefinition } from '../model-definition.js';
 import type { Attributes, Model, ModelStatic } from '../model.js';
 import { Op } from '../operators.js';
 import type { BindOrReplacements, Expression, Sequelize } from '../sequelize.js';
 import type { NormalizedOptions } from '../sequelize.types.js';
 import { bestGuessDataTypeOfVal } from '../sql-string.js';
-import { TableHints } from '../table-hints.js';
 import type { IsolationLevel } from '../transaction.js';
 import { rejectInvalidOptions } from '../utils/check.js';
 import { noOpCol } from '../utils/deprecations.js';
