+++ packages/core/src/abstract-dialect/query-interface-typescript.ts
@@ -3,9 +3,9 @@ import isEmpty from 'lodash/isEmpty';
 import assert from 'node:assert';
 import type { ConstraintChecking } from '../deferrable';
 import { Deferrable } from '../deferrable';
+import { QueryTypes } from '../enums';
 import { BaseError } from '../errors';
 import { setTransactionFromCls } from '../model-internals.js';
-import { QueryTypes } from '../query-types';
 import type { QueryRawOptions, QueryRawOptionsWithType, Sequelize } from '../sequelize';
 import { COMPLETES_TRANSACTION, Transaction } from '../transaction';
 import { isErrorWithStringCode } from '../utils/check.js';
